ACT-70: Reviewed coverage of Identity, Manegement and RepositoryService + detailed API coverage/docs of IdentityService

+++ modules/activiti-engine/src/main/java/org/activiti/engine/impl/cmd/DeleteUserCmd.java
@@ -12,6 +12,7 @@
  */
 package org.activiti.engine.impl.cmd;
 
+import org.activiti.engine.ActivitiException;
 import org.activiti.engine.impl.interceptor.Command;
 import org.activiti.engine.impl.interceptor.CommandContext;
 
@@ -28,6 +29,9 @@ public class DeleteUserCmd implements Command<Void> {
   }
 
   public Void execute(CommandContext commandContext) {
+    if(userId == null) {
+      throw new ActivitiException("userId is null");
+    }
     commandContext
       .getIdentitySession()
       .deleteUser(userId);
